Implement pattern-matching in "for" loops; fix pattern-matching semantics for sets.
This patch implements patterns in "for" statements:

for target in iter:
    body

such that 'target' can now be a pattern. Backward compatibility with
non-pattern "for" semantics is maintained by first checking whether
'target' contains any free variables (names starting with "_"), if
not, then fallback to non-pattern "for".

Also when pattern-matching over sets, the result is de-duplicated
after projection.

* dpy/compiler/parser.py:

  - (FreeVarFinder): New visitor.

  - (Parser.visit_For): If target contains freevars, treat it as a
    pattern.

* dpy/compiler/pygen.py:

  - (PythonGenerator.generate_pattern_domain): New method.

  - (PythonGenerator.visit_For): Handle ForStmt with patterns.

* dpy/patterns.py:

  - (PatternElement.filter): If iterable is a set, de-dup after
    projection.

  - (PatternElement._filter_list, PatternElement._filter_set): New
    methods.