Work around 'import_da' exception when using 'spawn' process creation method.
On 'spawn' platforms, the child process must reload modules into memory, which
means any module-level code will be run. This happens _before_ Process.run(),
where we do our child process initialization, resulting in an exception. This
is a temporary workaround that removes dependency on common.GlobalOptions for
'import_da'. The proper fix is to refactor the runtime code to de-couple the
generated user code from the real multiprocessing.Process class.

* da/api.py(import_da): just set `force_recompile` to False when we don't have
  common.GlobalOptions yet.