Comprehensions: remove 'has=' keyword, combine domains and conditions.
This patch changes the DistAlgo comprehensions so that the conditions
and domains can be interleaved, and the 'has=' keyword is removed. The
user can now simply specify comprehensions as an 'element' and a list of
conditions:

    setof(ELEM, COND1, ...)

At compilation time, a condition can be treated as either a
'domainspec', which must conform to a specific syntactical form and can
introduce new variable bindings, or a 'condition', which is any boolean
expression. A 'domainspec' is an expression of the form 'PATTERN in
DOMAIN', where 'PATTERN' is any tuple pattern, and is translated into
a Python comprehension: 'for TUPLE in DOMAIN if CONSTRAINTS'.

Evaluation of comprehensions follows simple left-to-right order, similar
to Python comprehensions, except in the case where conditions appear to
the left of the leftmost 'domainspec', in which case those conditions
are reordered to the right of the leftmost 'domainspec'. A DistAlgo
comprehension that contains no 'domainspec's degenerates to a Python
if-expression:

    {PYELEM} if (PYCOND1 and ...) else {}

* da/compiler/dast.py:

  -(QueryExpr): new expression type.

  -(ComprehensionExpr, QuantifiedExpr): extend 'QueryExpr'.

  -(ComprehensionExpr.domains): removed, combined into 'conditions'.

* da/compiler/parser.py:

  -(PatternParser): fix scoping rules: remove property 'out_scope', use
   'parser.current_scope' for all name lookups.

  -(Pattern2Constant): signal error if 'FreePattern' encountered.

  -(Parser.current_query, Parser.current_query_scope): now properties
   inferred from the node stack, instead of concrete values.

  -(Parser.leave_query): remove redundant 'node' argument, since current
   node is always top of node stack; new argument 'merge', if True,
   merges current scope into the query scope; new argument 'audit', if
   True, performs query auditing if current node is the top-level query.

  -(Parser.parse_comprehension): update to parse new DistAlgo
   comprehension style.

* da/compiler/pygen.py:

  -(propagate_attributes): allow argument 'from_nodes' to be single
   node.

  -(propagate_subexprs): not used, removed.

  -(propagate_fields): new function, given a 'node', propagates
   attributes from its children (as specified by 'node._fields').

  -(PythonGenerator.visit_ComprehensionExpr): update to support new
   comprehension style.

  -(PythonGenerator.visit_DomainSpec): changed traversal order: visit
   domain first.

* examples/*: update to new comprehension style.