Instrument 'freeze' for jbstyle.
Invinc needs to keep track of all changes to sets and maps used by the
system, so we can not simply 'freeze' them without telling Invinc. This
patch changes 'freeze' so that it deep-copies instead of freezing when
it detects Invinc sets and maps. This should be sufficient for our
examples, since Invinc set and map are hashable, and is probably the
best we can do until Invinc gets proper frozen set and map.

* da/common.py:

  - (InvIncBaseType): new global var, set to "invinc.runtime.Type" when
    using Invinc, None otherwise.

  - (get_inc_module): new function.

  - (load_modules): sets 'InvIncBaseType' when IncModule.JbStyle is set.

  - (freeze): deepcopy Invinc objects instead of freezing.

* da/compiler/incgen.py:

  - (JB_STYLE_MODULE, JB_STYLE_SET, JB_STYLE_MAP): new constants.

  - (PREAMBLE): changed to be more descriptive; add 'JbStyle' global
    boolean value.

  - (generate_update_stub): always use 'return' if the update is an
    expression, regardless of 'jb_style'.

  - (process_initializers): use new JB_STYLE_* constants instead of
    hardcoded strings.

  - (generate_header): adjust for new PREAMBLE format.