incgen: make initializers run in the child process instead of the parent.
Previously, incgen injected initializer calls in the '__init__' method
of process classes, which means the initializers were run in the parent
process address space. This was fine when 'fork' semantics were used for
child process creation. However, when using 'spawn' semantics (i.e.
Windows) the inc module namespace would not be initalized in the child
process, causing an error. This patch moves all initializer calls to the
'setup' method so they'll be called from the child process instead.

* da/compiler/incgen.py:

  - (PREAMBLE): remove 'init' function definiton.

  - (generate_event_initializer_stub): removed. This function generated
    assignment stubs for event histories which was never used.

  - (process_events): remove call to 'generate_event_initializer_stub'.

  - (INIT_FUN): new constant containing definiton of 'init'
    function (moved from 'PREAMBLE').

  - (process_initializers): generate the 'init' function, and inject
    calls to initialize jbstyle sets and maps.

  - (process_setups): inject calls to event history initializers.

  - (StubcallGenerator.history_initializers): don't generate anything
    here since history initializers are moved to 'setup'.