incgen: implement witness for top-level existential quantifications.
The witness is obtained by calling "next(iter())" on a "witness set".
The "witness set" for a given query is determined by the transformation
method (tables 1, 2, 3, or plain), and contains bindings for the free
variables of the top-level query. The "witness set" is part of the full
query and must be repeated in the computation of the query
result (instead of assigning to a temp variable) in order for invinc to
recognize them as one query. We use a global var in the inc module (name
given by 'GLOBAL_WITNESS_VAR') to temporarily store a witness value for
the last existential quantification query. In the main module, an inner
function is generated to wrap the call to the query stub, and after the
query returns, the witness value is propagated through nonlocal
bindings.

* da/compiler/dast.py(QuantifiedExpr.ordered_local_freevars): new
  property, returns ordered list of free variables in the domain specs
  of this QuantifiedExpr.

* da/compiler/incgen.py:

  - (LOCAL_WITNESS_SET, LOCAL_RESULT_VAR, GLOBAL_WITNESS_VAR): new constants.

  - (PREAMBLE, generate_header): changed to use GLOBAL_WITNESS_VAR.

  - (process_query): refactored into a witness case and a non-witness case.

  - (_process_nonwitness_query, _process_witness_query): new functions.

  - (QueryExtractor): don't lump QuantifiedExpr inside BooleanExpr into
    the same query, instead treat each top-level QuantifiedExpr as a
    separate query.

  - (IncInterfaceGenerator.witness_set): new attribute, used to store
    the witness set when one is needed.

  - (IncInterfaceGenerator.do_table1_transformation,
    IncInterfaceGenerator.do_table2_transformation,
    IncInterfaceGenerator.do_table3_transformation): extract witness set
    for top-level existential quantifications.