pygen: fix scoping bug in nested quantifications.
Quantifications are translated to functions. When they are nested in
another query, any names bound by the outer queries must be passed in as
arguments (since they are not visible to the body of the function).
Previously, names that are defined in a higher scope than the top-level
query were not correctly identified as parameters of the inner
quantification. We fix this by changing the parameter accumulation logic
to walk upward in the AST and collect all free variables in the
containing queries, instead of relying on the scope information in the
name objects.

* da/compiler/dast.py:

  - (QueryExpr.ordered_local_freevars): new abstract property.

  - (ComprehensionExpr.ordered_local_freevars): new property, like
    'QuantifiedExpr.ordered_local_freevars'.

* da/compiler/pygen.py(PythonGenerator.visit_QuantifiedExpr): do an
  upward tree walk to collect 'params', instead of using scope info in
  name objects.