Lapaxos: fix long time bug in Acceptor.
Wrong element expression was used in the "max(setof)" query in the
handler for 'prepare' messages, causing the "max proposed val" field in
the Acceptor's 'respond' message to always be None. This in term causes
each Proposer to always choose a random value for its 'accept' message.
Interestingly, the learners will still be able to reach a consensus, due
to the "max responded ballot number check" on line 45.