Parser: don't require 'import' statements to be at top-level.
This restriction was originally put in place because we wanted to hook
into the import mechanism to do auto-compilation of da files. That plan
fell apart since there's no easy way to accomplish this with the way the
Python import system was designed. Importing 'da' files now requires the
programmer to use the 'da_import' API, so this restriction is no longer
required.