Updated lapaxos and sktoken benchmark scripts regarding timeout and list length.
Made timeout a command-line argument and setup parameter in cases where
it was previously a hard-coded constant. Eliminated an arbitrary
multiplier for the timeout. Replaced len(<list>) with an alias for len
so the inc interface wouldn't treat it as a query (only queries with len
of sets are supported in incoq).