Work around firewall problems on Windows.
On non-trusted networks, Windows firewall blocks all traffic to sockets
not explicitly listening on the loop-back interface, even if the traffic
originated from the local machine. The symptom is that DistAlgo programs
run fine when the system is offline, but would hang when connected to a
network marked as non-trusted.

As a work around, we now bind to the loop-back interface by default,
instead of the interface corresponding to the system hostname. This
prevents DistAlgo processes from communicating with processes running on
remote hosts unless the user explicitly specifies another interface for
new processes to bind to. To have DistAlgo processes bind to a different
interface, specify it using the 'name' keyword argument to 'new'.

* da/endpoint.py(EndPoint.__init__): set default hostname to
  'localhost'.