Fix last fix for ramutex/spec.
To compute value of `ownc' in P.receive('request'...), it is not
sufficient to just use the max of our 'request's, we have to also check
that we haven't already released. This is tricky since there is no
'release' message in ramutex. Instead, we look for 'ack' from all
peers (that is greater than the request). Due to the atomicity of the
critical section, this implies we either have released or are releasing
the mutex corresponding to the request.

* examples/ramutex/spec.da, benchmarks/ramutex/spec.da:
  - (P.cs): move 'cs' label to the await statement, so that the critical
  section and the satisfaction of the await condition is one atomic
  block.
  - (P.receive): add condition to filter out the 'request' messages that
  have been 'ack'ed by everybody.