Add support for passing verbatim code to inc interface.
This patch introduces a new keyword, `_INC_', to pass verbatim code to
the inc interface file. `_INC_' must be used at the module level, and is
of the form:

    _INC_("""
        CODE
    """)

The contents of 'CODE' must be valid Python code, and will be included
verbatim in the generated inc interface file.

This mechanism is useful for passing program-specific hints to the
incrementalizer, such as variable type information.

* da/compiler/dast.py(Program.directives): new node attribute.

* da/compiler/incgen.py(generate_header): include `input_ast.directives'
  after the import statements, and before any global variable
  definitions.

* da/compiler/parser.py:
  - (KW_INC_VERB): new keyword, only has affect at the module level.
  - (parser.visit_Expr): parse inc directives.