Use `deque' instead of `queue' for event queue.
The `queue.Queue' is a convenient thread-safe, waitable queue
implementation, but has extremely high overhead due to locking. The
`collections.deque' class has similar interface to `Queue', and its
push/pop operations are around 40 times faster than the equivalent
operations on `Queue'. A nice side-effect of the GIL is that built-in
operations such as push/pop on the `deque' are automatically
thread-safe, so we can _almost_ use `deque' as a drop-in replacement for
`Queue'. The only missing piece is a blocking wait for the "consumer"
when the `deque' is empty, and that can be easily implemented using a
condition variable.

This gives us more precise control over locking. Specifically, we only
need to lock if the the `block' argument is set, the remaining timeout
is greater than zero, _and_ the queue is empty. In all other cases we
can completely avoid the lock and the associated trap to kernel.

* da/sim.py:
  -(Comm.__init__): move all Comm related state initialization here: set
   'daemon'; create the condition var and the deque; initialize
   `num_waiting' to 0, this counter is used to support multiple DistAlgo
   user threads.
  -(Comm.run): use `deque'; notify the condition variable if there are
   threads waiting.
  -(DistProcess.__init__): `_jobqueue' doesn't need to be a `deque', a
   list works fine.
  -(DistProcess._start_comm_thread): let Comm do its own initialization.
  -(DistProcess._label_one): remove redundant 'nmax'.
  -(DistProcess._label_all): simplify logic.
  -(DistProcess._process_jobqueue): simplify logic.
  -(DistProcess._process_event): re-implemented for deque+condition.