Refactor: wrap program `main` in a `DistProcess`.
Currently, the DistAlgo program entry point(`main()`) is run as a normal
Python function inside the initial Python process. This limits the
ability of `main` to interact with DistAlgo processes -- we have a hack
that made it possible to `send` messages from `main`, but it can neither
receive messages nor use `await` to synchronize.

This patch makes the DistAlgo system run the `main` method inside a
full-blown DistAlgo process, so that the full range of DistAlgo
primitives is now available inside `main`. This patch also removes the
last remaining global states from the `api` module, making `api` fully
re-entrant.

* da/api.py:
  -(RootLock): removed; was used for synchronizing the collection of
   performance data from child processes.
  -(EndPointType, get_channel_type, send): removed; was used to create a
   dummy process id for sending messages from the `main` function.
  -(entrypoint): create and run the `_NodeMain` process instead of
   running the `main` function directly.

* da/compiler/pygen.py:
  -(NODECLS_NAME, ENTRYPOINT_NAME): new global constants.
  -(PythonGenerator.visit_Program): call `generate_nodecls` to generate
   a process class to wrap the `main` function.
  -(PythonGenerator.generate_nodecls): generate a process class to wrap the
   `main` function.
  -(PythonGenerator._entry_point): force the name and arguments of entry
   point functions to be `run(self)` so we can use this method for both
   process entrypoints(`run`) and program entrypoints(`main`).

* examples/lapaxos/orig.da, examples/vrpaxos/orig.da,
  examples/vrpaxos/spec.da, benchmarks/vrpaxos/orig.da: use `send`
  instead of `da.send` in the `main` functions.