Reverse the use of '@' and '#' in process names.
We will try stick with the Erlang naming convention as much as possible,
so a node name would have the form "node[@host[:port]]", and a process
name would have the form "name[#NODENAME]".

The question remains what would be the best way to partition the global
process name space. 74de4c5a7ce743766c0da981dcf76dabd23bd5f1 gave each
node its own name space, so "a#b" and "a#c" refer to two different
processes both named "a", where one is running on node "b" and the other
is running on node "c". But the node name space is still globally flat,
which means that having "b@h" and "b@i" would be a name conflict. Thus
it seems desirable to further partition the node name space by
hostnames. The problem with that is 1. we don't actually keep track of
hostnames in process ids, only IP addresses; 2. there maybe multiple
hostnames all resolving to the same IP.

* da/common.py:
  -(name_split_host): use '@' as separator.
  -(name_split_node): use '#' as separator.