Fix inheritance: call parent `setup` from sub-class.
This patch fixes process-class inheritance so that the `setup()` methods
defined in parent classes are properly invoked.

The reason this wasn't done before was due to the fact that the DistAlgo
compiler operates only on local information inside the module being
compiled -- it does not have access to type and signature information on
classes and methods defined in other modules, which is often the case
with parent classes, and without the signature of the parent class'
`setup` method so it doesn't really know how to properly call it. We
side-step this issue by using Python's keyword argument, which basically
allows passing arbitrary arguments to functions.

* da/compiler/parser.py(Parser.visit_ClassDef): disallow the user to
  define the keyword argument for their `setup` method, since we need
  that to pass arguments to the parent `setup`.

* da/compiler/pygen.py:
  -(PATTERN_EXPR_NAME, QUATIFIED_EXPR_NAME): removed because not used.
  -(CATCHALL_PARAM_NAME): new global constant.
  -(pyCall, pyClassDef): move the version checks inside, the runtime
  overhead for the checks is negligible.
  -(PythonGenerator):
    -(body): new parameter 'res', allows caller to pass a list object
    which we then append to, this avoids unnecessary copying when
    statements needs to be inserted at the front of a block.
    -(visit_Arguments): be more precise with the version check.
    -(_generate_setup): new function, refactored from `_visit_Function`.
    -(visit_Function): call `_generate_setup`; pass 'fd.body' to
    `self.body`.

* da/sim.py(DistProcess.setup): add a keyword argument to catch anything
  that may be passed as argument when the sub-class `setup` calls
  `super().setup`.