send: support sender impersonation.
It may sometimes be necessary to pretend the sender is someone other
than 'self'. For example, in classes that does message packing and
unpacking, a common paradigm is to unpack an incoming message and send
it back to 'self' for further processing:

    def receive(msg=(_TAG, message)):
        # do something about TAG
        super().send(message, to=self)

This kind of handlers are usually defined in super-classes, which
provide services such as encryption and decryption. This allows
separation of the program logic into "layered" component, and composing
them simply through process-class inheritance. The only issue is that
when sending the unpacked message back to 'self', the sender now becomes
'self', so this composition is not entirely transparent to child
classes.

To address this we add an optional (and hidden) keyword 'impersonate',
which allows the sender to pretend the message came from someone else.
The above handler now becomes:

    def receive(msg=(_TAG, message), from_=realsrc):
        # do something about TAG
        super().send(message, to=self, impersonate=realsrc)

which makes unpacked message appear to come from the original sender.
And thus the composition can be made fully transparent to child classes.

* da/sim.py:
  -(DistProcess):
    -(send): pass 'impersonate' to '_send1'.
    -(_send1): if 'impersonate' is not None, use it as the src of the
    message, otherwise, default to `self.id`.