Fix potential infinite loop in router.
When routing a message to a process id that is not in Router.local_procs but
has the same transport address as the local TransportManager (this will
happen when e.g. the destination process has already terminated), if the
router forwards the message to the TransportManager it will come right back
through Router.incomingq, causing an infinite loop. This patch prevents this
by discarding messages that will cause infinite circular routing.

* da/sim.py:
  -(CircularRoutingException): new exception class.
  -(Router):
   -(._send_remote): check whether destination is actually remote (i.e. has
    different transport address as local TransportManager)
   -(._dispatch): catch and log CircularRoutingException.