Change default value of '--hostname'.
When the user specifies '--nodename', he/she likely intends to run the node
as part of a distributed system, in which case defaulting '--hostname' to
'localhost' would be a poor choice because it precludes distributed
communication. The fully qualified domain name of the host system, as
returned by `socket.getfqdn()`, would serve as a much better default that
allows distributed communication, assuming that it is properly configured in
the OS(i.e. '/etc/hostname' and '/etc/hosts' on *nix systems).

* da/__main__.py(parseArgs): update description of '--hostname'.

* da/api.py(_bootstrap_node): improve log message for bootstrap failure.

* da/common.py(initialize_runtime_options): set default value of
  '--hostname' according to value of '--nodename'.