Add files via upload
daast_nest.py -- defines a visitor for the DistAlgo AST, which prints out the structure of the AST by printing the class name for each AST node it visits. Child node names are indented relative to their parents, giving a clear presentation of how DistNodes in the tree are related to eachother.

daast_dict.py -- defines a visitor for the DistAlgo AST, which prints out the contents--the attribute names and attribute values for each DistNode that it visits (using the Python __dict__ attribute).

Together the visitors output a clear representation of both the gross and fine structure of the DistAlgo AST for a given DistAlgo source file.