Fix inheritance cont.: allow declaring new names in `setup()`.
(This patch is a continuation of 54f7f950470afea848352ab9573ea4fb520d9f9d.)
Currently, there's no way for a child process-class to use state variables
defined in a parent process-class, due to the lack of name resolution. This
patch introduces a band-aid fix for this problem, by allowing the explicit
declaration of process state variables names in `setup()` methods, so that a
programmer can manually introduce parent class variable names to the child
class, like so:

    class Parent(process):
        def setup():
            self.var_a = 1
        ...

    class Child(Parent, process):
        def setup():
            self.var_a     # declares `var_a` from the parent class

        def run():
            output(var_a)  # will print '1'

Note that if a name thus declared was not defined in any parent classes,
then a runtime error will result.

This issue has been properly addressed on the master branch, commit
521004c40be8525720881aa23c3b12d9e2d14c0e, but will not be back-ported to
1.0.z.

* da/compiler/parser.py(Parser.visit_Expr): if we're in `setup`, set context
  to `Assignment` so that new names can be added.