TCP: use blocking mode for client-side sockets.
We used to set a timeout of 5 seconds on TCP client sockets, mainly so that
we would fail faster if the server side is down or blocked by some firewall.
The downside is that this puts the socket in non-blocking mode, and
`socket.sendmsg` does not guarantee to send the whole message before
returning for non-blocking sockets. To fix this problem, we would either
have to add logic to `TcpTransport._send_1` to handle partial sends, or
simply use blocking mode and let the system handle the send. This patch
implements the latter approach, since bailing out a few seconds faster on a
non-responsive server is not worth complicating the more frequently used
`_send_1`.

* da/endpoint.py(TcpTransport._connect): set socket timeout to None (i.e.
  blocking mode).