Re-resolve process name if user-specified hostname differs from cached.
When resolving a process name with an '@' suffix, if the user-specified
hostname is different from the one in the cached ProcessId, it might be that
the user has moved the process name to a new host, so we should retry
name-resolution using the user-specified hostname. Even if the
'@' suffix is actually stale, i.e. the user-specified hostname is down or
running an *older* process, it wouldn't matter as we ensure newer ProcessIds
take precedence over older ones.

* da/common.py(name_split_host): canonize hostname.

* da/sim.py(DistProcess.resolve): re-resolve name if the cached ProcessId's
  hostname is different from the user-specified suffix.