Incorporate hostname into process name uniqueness test.
This is a followup patch to 23ed359d776064e41988cd8dd8c1766d96f49224. Ever
since the introduction of the '@host' suffix by
74de4c5a7ce743766c0da981dcf76dabd23bd5f1, there has been the problem of
whether process names such as 'A@M' and 'A@N' should be treated as the same
or different names. Our original design was to provide a flat process name
space that abstracts away the physical hosts they run on, so that 'A@M' and
'A@N' both refer to the same process 'A', which may have migrated from one
host to another. The introduction of the '@host' suffix broke this
abstraction, leading to issues described in
23ed359d776064e41988cd8dd8c1766d96f49224. Unfortunately,
23ed359d776064e41988cd8dd8c1766d96f49224 was not an adequate fix for the
issue, and taking that approach further would just needlessly complicate
and slow down name resolution.

This patch adopts a "full measure" -- we abandon our original design of a
flat process name space that hides physical hosts, and simply incorporate
hostname into the full process name. If the hostname is not specified(i.e.
None), we default to the host the current process is running on.

* da/common.py(ProcessId):
  -(__new__): incorporate hostname into fullname.
  -(lookup_or_register_callback): save a dictionary lookup.

* da/sim.py:
  -(DistProcess.resolve): use current host as default; incorporate host into
   full name.
  -(NodeProcess.bootstrap): downgrade bootstrap failure message from error
   to warning.
  -(NodeProcess._cmd_Resolve): adapt to new `args` tuple format.